+++ plugin-packs/postcss-bundler/src/postcss-import/lib/parse-styles.ts
@@ -1,8 +1,8 @@
 import type { Document, Postcss, Result, Root, AtRule } from 'postcss';
-import { CharsetStatement, ImportStatement, Statement, isCharsetStatement, isImportStatement } from './statement';
+import { ImportStatement, Stylesheet, isImportStatement } from './statement';
 import { Condition } from './conditions';
 import { isValidDataURL } from './data-url';
-import { parseStatements } from './parse-statements';
+import { parseStylesheet } from './parse-stylesheet';
 import { createRequire, resolveId } from './resolve-id';
 import { loadContent } from './load-content';
 import noopPlugin from './noop-plugin';
@@ -15,8 +15,8 @@ export async function parseStyles(
 	conditions: Array<Condition>,
 	from: Array<string>,
 	postcss: Postcss,
-) {
-	const statements = parseStatements(result, styles, importingNode, conditions, from);
+): Promise<Stylesheet> {
+	const stylesheet = parseStylesheet(result, styles, importingNode, conditions, from);
 
 	{
 		// Lazy because the current stylesheet might not contain any further @import statements
@@ -25,7 +25,7 @@ export async function parseStyles(
 		let base: string | undefined;
 
 		const jobs: Array<Promise<void>> = [];
-		for (const stmt of statements) {
+		for (const stmt of stylesheet.statements) {
 			if (!isImportStatement(stmt) || !isProcessableURL(stmt.uri)) {
 				continue;
 			}
@@ -40,42 +40,37 @@ export async function parseStyles(
 			jobs.push(resolveImportId(result, stmt, postcss, require, sourceFile, base));
 		}
 
-		await Promise.all(jobs);
+		if (jobs.length) {
+			await Promise.all(jobs);
+		}
 	}
 
-	let charset: CharsetStatement | null = null;
-	const imports: Array<Statement> = [];
-	const bundle: Array<Statement> = [];
+	for (let i = 0; i < stylesheet.statements.length; i++) {
+		const stmt = stylesheet.statements[i];
 
-	// squash statements and their children
-	statements.forEach(stmt => {
-		if (isCharsetStatement(stmt)) {
-			charset = handleCharset(stmt, charset);
-		} else if (isImportStatement(stmt)) {
-			if (stmt.children) {
-				stmt.children.forEach((child) => {
-					if (isImportStatement(child)) {
-						imports.push(child);
-					} else if (isCharsetStatement(child)) {
-						charset = handleCharset(child, charset);
-					} else {
-						bundle.push(child);
-					}
-				});
-			} else {
-				imports.push(stmt);
+		if (isImportStatement(stmt) && stmt.stylesheet) {
+			if (stylesheet.charset && stmt.stylesheet.charset && stylesheet.charset.params.toLowerCase() !== stmt.stylesheet.charset.params.toLowerCase()) {
+				throw stmt.stylesheet.charset.error(
+					'Incompatible @charset statements:\n' +
+					`  ${stmt.stylesheet.charset.params} specified in ${stmt.stylesheet.charset.source?.input.file}\n` +
+					`  ${stylesheet.charset.params} specified in ${stylesheet.charset.source?.input.file}`,
+				);
+			} else if (!stylesheet.charset && !!stmt.stylesheet.charset) {
+				stylesheet.charset = stmt.stylesheet.charset;
 			}
-		} else if (stmt.type === 'nodes') {
-			bundle.push(stmt);
-		}
-	});
 
-	return charset ? [charset, ...imports.concat(bundle)] : imports.concat(bundle);
+			stylesheet.statements.splice(i, 1, ...stmt.stylesheet.statements);
+			i--;
+			continue;
+		}
+	}
+
+	return stylesheet;
 }
 
 async function resolveImportId(result: Result, stmt: ImportStatement, postcss: Postcss, require: NodeRequire, sourceFile: string, base: string) {
 	if (isValidDataURL(stmt.uri)) {
-		stmt.children = await loadImportContent(
+		stmt.stylesheet = await loadImportContent(
 			result,
 			stmt,
 			stmt.uri,
@@ -86,7 +81,7 @@ async function resolveImportId(result: Result, stmt: ImportStatement, postcss: P
 	} else if (isValidDataURL(stmt.from[stmt.from.length - 1])) {
 		// Data urls can't be used as a base url to resolve imports.
 		// Skip inlining and warn.
-		stmt.children = [];
+		stmt.stylesheet = { statements: [] };
 		result.warn(
 			`Unable to import '${stmt.uri}' from a stylesheet that is embedded in a data url`,
 			{
@@ -105,8 +100,7 @@ async function resolveImportId(result: Result, stmt: ImportStatement, postcss: P
 		parent: sourceFile,
 	});
 
-	const importedContent = await loadImportContent(result, stmt, resolved, postcss);
-	stmt.children = importedContent ?? [];
+	stmt.stylesheet = await loadImportContent(result, stmt, resolved, postcss);
 }
 
 async function loadImportContent(
@@ -114,10 +108,10 @@ async function loadImportContent(
 	stmt: ImportStatement,
 	filename: string,
 	postcss: Postcss,
-) {
+): Promise<Stylesheet> {
 	const { conditions, from, node } = stmt;
 	if (from.includes(filename)) {
-		return;
+		return { statements: [] };
 	}
 
 	let content: string;
@@ -177,20 +171,4 @@ function isProcessableURL(uri: string): boolean {
 	}
 
 	return true;
-}
-
-function handleCharset(stmt: CharsetStatement, ref: CharsetStatement | null = null): CharsetStatement {
-	if (!ref) {
-		return stmt;
-	} else if (
-		stmt.node.params.toLowerCase() !== ref.node.params.toLowerCase()
-	) {
-		throw stmt.node.error(
-			`Incompatible @charset statements:
-  ${stmt.node.params} specified in ${stmt.node.source?.input.file}
-  ${ref.node.params} specified in ${ref.node.source?.input.file}`,
-		);
-	}
-
-	return ref;
 }